requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.0.0
Django==4.2.3
django-cleanup==8.0.0
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-parler==2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
geographiclib==2.0
geopy==2.4.1
gunicorn==20.1.0
idna==3.6
Pillow==10.0.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dateutil==2.9.0.post0
pytz==2023.3
requests==2.31.0
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.2.1
whitenoise==6.5.0
